+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [x]
+- Config reporting view per object
+- Default user reporting view
+
+## [2.5.0]
+- Enable lp for folders
+- Org unit filter config
+
 ## [2.4.0]
 - Course management:
     - User language column and filter